cdc: reduce events batch to solve congest error (#11086)
* cdc: reduce events batch to save memory


* cdc: remove histogram in hot path
