+++ components/cdc/src/lib.rs
@@ -13,9 +13,9 @@ mod observer;
 mod old_value;
 mod service;
 
-pub use channel::{recv_timeout, MemoryQuota};
+pub use channel::{recv_timeout, CdcEvent, MemoryQuota};
 pub use config::CdcConfigManager;
 pub use endpoint::{CdcTxnExtraScheduler, Endpoint, Task, Validate};
 pub use errors::{Error, Result};
 pub use observer::CdcObserver;
-pub use service::{CdcEvent, FeatureGate, Service};
+pub use service::{FeatureGate, Service};
